Fix #10229 Idle Timeout
Do not use Wrappers for the wrapped servlet request/response
